Update README.md
Stranger Things is a TV series about a group of kids in the 1980s who encounter supernatural events in the fictional town of Hawkins, Indiana: 
Setting: The show is set in the 1980s in the fictional town of Hawkins, Indiana, where the nearby Hawkins National Laboratory conducts secret paranormal experiments. 
Plot: The show follows the story of a group of kids as they encounter strange supernatural events that are the result of a secret science experiment gone wrong. The first season begins when Will Byers is abducted by a creature from the Upside Down, an alternate dimension that reflects the real world. 
Characters: The show stars Millie Bobby Brown, Finn Wolfhard, and Winona Ryder. 
Creators: The show was created by the Duffer Brothers, Matt and Ross Duffer. 
Rating: The show has a TV-14 rating and is a drama, fantasy, and horror show. 
Awards: The show has won 12 Primetime Emmys.